Remove the tight-coupling to Eclipse.
The branch maven-restructure-dev now will work on IDEA, Eclipse, or any other IDEA that has maven integration with just one click!

Need to remove the binary files such as .eclipse, .project, and .jar that are committed into the source. But that is of a less priority now.

Next task is to build and ensure that nothing is broken.