First commit on #39.
Removed embedded jars from drill-datasource-provider.

Tested and confirmed that everything is working fine, and Bindaas drill provider is functional.

Need to repeat this process to every single module in Bindaas.

Ideally, we will have more than 90% in the source file zip since we are going to remove most, if not all, embedded jars. At each step of removing jar, I confirm that Bindaas is not broken